requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
pandas>=2.0.0
openpyxl>=3.1.0
